Added short and long variants for BotW bingo
