+++ bingosync-app/bingosync/game_type.py
@@ -94,6 +94,8 @@ class GameType(Enum):
     yooka_laylee = 88
     ocarina_of_time_item_randomizer = 89
     ocarina_of_time_item_randomizer_blackout = 90
+    zelda_botw_short = 91
+    zelda_botw_long = 92
 
     def __str__(self):
         return self.short_name
@@ -282,6 +284,14 @@ GAME_GROUPS = {
             (GameType.resident_evil_hd_randomizer, "Randomizer", "REHD Random"),
         ],
     },
+    GameType.zelda_botw: {
+        "name": "Zelda: Breath of the Wild",
+        "variants": [
+            (GameType.zelda_botw, "Normal", "BotW Normal"),
+            (GameType.zelda_botw_short, "Short", "BotW Short"),
+            (GameType.zelda_botw_long, "Long", "BotW Long"),
+        ],
+    },
     **singleton_group(GameType.super_mario_64, "Super Mario 64", "SM64"),
     **singleton_group(GameType.majoras_mask, "Zelda: Majora's Mask", "Zelda: MM"),
     **singleton_group(GameType.super_metroid, "Super Metroid", "Super Metroid"),
@@ -303,7 +313,6 @@ GAME_GROUPS = {
     **singleton_group(GameType.psychonauts, "Psychonauts", "Psychonauts"),
     **singleton_group(GameType.super_mario_galaxy, "Super Mario Galaxy", "SM Galaxy"),
     **singleton_group(GameType.banjo_tooie, "Banjo-Tooie", "Banjo-Tooie"),
-    **singleton_group(GameType.zelda_botw, "Zelda: Breath of the Wild", "Zelda: BotW"),
     **singleton_group(GameType.the_witness, "The Witness", "The Witness"),
     **singleton_group(GameType.pikmin_2, "Pikmin 2", "Pikmin 2"),
     **singleton_group(GameType.pokemon_platinum, "Pokémon Platinum", "Poké Plat."),
