+++ bingosync-app/testdata/gen_output/zelda_botw_long/1.json
@@ -1,0 +1,77 @@
+[
+    {
+        "name": "'The Giant of Ralis Pond' Side Quest"
+    },
+    {
+        "name": "'Divine Beast Vah Naboris' Main Quest"
+    },
+    {
+        "name": "Take a Selfie while on fire"
+    },
+    {
+        "name": "15 Akkala Koroks"
+    },
+    {
+        "name": "11 Shrines"
+    },
+    {
+        "name": "6 Shrine Quests"
+    },
+    {
+        "name": "10 Creatures in Compendium"
+    },
+    {
+        "name": "'Frog Catching' Side Quest"
+    },
+    {
+        "name": "Activate 10 Towers"
+    },
+    {
+        "name": "25 Koroks"
+    },
+    {
+        "name": "45 Koroks"
+    },
+    {
+        "name": "Snowquill Tunic \u2605"
+    },
+    {
+        "name": "1 Divine Beast Main Quest"
+    },
+    {
+        "name": "11 Shrines"
+    },
+    {
+        "name": "Collect 5 Picture Memories"
+    },
+    {
+        "name": "10 Creatures in Compendium"
+    },
+    {
+        "name": "14 Side Quests"
+    },
+    {
+        "name": "Climbing Gear \u2605"
+    },
+    {
+        "name": "Kill a Guardian Skywatcher"
+    },
+    {
+        "name": "'Divine Beast Vah Ruta' Main Quest"
+    },
+    {
+        "name": "Flamebreaker Helm"
+    },
+    {
+        "name": "Kill Coliseum Lynel"
+    },
+    {
+        "name": "'The Spring of Wisdom' Shrine Quest"
+    },
+    {
+        "name": "20 Creatures in Compendium"
+    },
+    {
+        "name": "Activate 10 Towers"
+    }
+]
