+++ bingosync-app/testdata/gen_output/zelda_botw_long/1000.json
@@ -1,0 +1,77 @@
+[
+    {
+        "name": "50 Koroks"
+    },
+    {
+        "name": "'Luminous Stone Gathering' Side Quest"
+    },
+    {
+        "name": "6 Shrine Quests"
+    },
+    {
+        "name": "10 Creatures in Compendium"
+    },
+    {
+        "name": "Finish Basic Shield Surfing Minigame"
+    },
+    {
+        "name": "Snow Boots \u2605"
+    },
+    {
+        "name": "10 Shrines"
+    },
+    {
+        "name": "Activate 8 Towers"
+    },
+    {
+        "name": "'Riddles of Hyrule' Side Quest"
+    },
+    {
+        "name": "15 Tabantha Koroks"
+    },
+    {
+        "name": "'The Ceremonial Song' Shrine Quest"
+    },
+    {
+        "name": "20 Equipment in Compendium"
+    },
+    {
+        "name": "1 Fire, 1 Ice, and 1 Electric weapons"
+    },
+    {
+        "name": "Kill a Molduga"
+    },
+    {
+        "name": "Activate 9 Towers"
+    },
+    {
+        "name": "Activate 4 Towers"
+    },
+    {
+        "name": "Full Barbarian Set"
+    },
+    {
+        "name": "Collect 5 Picture Memories"
+    },
+    {
+        "name": "Mount the Lord of the Mountain"
+    },
+    {
+        "name": "11 Shrines"
+    },
+    {
+        "name": "1 Divine Beast Main Quest"
+    },
+    {
+        "name": "15 Dueling Peaks Koroks"
+    },
+    {
+        "name": "Sapphire Circlet"
+    },
+    {
+        "name": "Beat Ganon"
+    },
+    {
+        "name": "Collect 6 Picture Memories"
+    }
+]
