+++ bingosync-app/testdata/gen_output/zelda_botw_long/1234.json
@@ -1,0 +1,77 @@
+[
+    {
+        "name": "Zora Helm \u2605"
+    },
+    {
+        "name": "'Divine Beast Vah Rudania' Main Quest"
+    },
+    {
+        "name": "20 Equipment in Compendium"
+    },
+    {
+        "name": "14 Side Quests"
+    },
+    {
+        "name": "12 Shrines"
+    },
+    {
+        "name": "'Riddles of Hyrule' Side Quest"
+    },
+    {
+        "name": "11 Shrines"
+    },
+    {
+        "name": "12 Wasteland Koroks"
+    },
+    {
+        "name": "Beat Ganon"
+    },
+    {
+        "name": "Collect 3 Picture Memories"
+    },
+    {
+        "name": "2 Divine Beast Main Quest"
+    },
+    {
+        "name": "Kill Coliseum Lynel"
+    },
+    {
+        "name": "Spring-Loaded Hammer"
+    },
+    {
+        "name": "Activate 5 Towers"
+    },
+    {
+        "name": "Obtain a Material from Naydra"
+    },
+    {
+        "name": "10 Creatures in Compendium"
+    },
+    {
+        "name": "12 Wasteland Koroks"
+    },
+    {
+        "name": "Full Climber's Set"
+    },
+    {
+        "name": "'Divine Beast Vah Ruta' Main Quest"
+    },
+    {
+        "name": "Spring-Loaded Hammer"
+    },
+    {
+        "name": "Activate 6 Towers"
+    },
+    {
+        "name": "'Face the Frost Talus' Side Quest"
+    },
+    {
+        "name": "12 Side Quests"
+    },
+    {
+        "name": "'Test of Will' Shrine Quest"
+    },
+    {
+        "name": "50 Koroks"
+    }
+]
