+++ bingosync-app/testdata/gen_output/zelda_botw_long/12345.json
@@ -1,0 +1,77 @@
+[
+    {
+        "name": "'Frog Catching' Side Quest"
+    },
+    {
+        "name": "'Divine Beast Vah Medoh' Main Quest"
+    },
+    {
+        "name": "'Test of Will' Shrine Quest"
+    },
+    {
+        "name": "Mount the Lord of the Mountain"
+    },
+    {
+        "name": "Full Climber's Set"
+    },
+    {
+        "name": "10 Creatures in Compendium"
+    },
+    {
+        "name": "Full Barbarian Set"
+    },
+    {
+        "name": "14 Side Quests"
+    },
+    {
+        "name": "15 Hebra Koroks"
+    },
+    {
+        "name": "Kill a Guardian Skywatcher"
+    },
+    {
+        "name": "Climbing Gear"
+    },
+    {
+        "name": "'Death Mountain's Secret' Side Quest"
+    },
+    {
+        "name": "10 Creatures in Compendium"
+    },
+    {
+        "name": "15 Shrines"
+    },
+    {
+        "name": "Remake a Champion's Weapon or Shield"
+    },
+    {
+        "name": "Urbosa's Fury"
+    },
+    {
+        "name": "'The Ceremonial Song' Shrine Quest"
+    },
+    {
+        "name": "15 Hebra Koroks"
+    },
+    {
+        "name": "Kill a Black Hinox"
+    },
+    {
+        "name": "Activate 5 Towers"
+    },
+    {
+        "name": "Kill a Lynel"
+    },
+    {
+        "name": "1 Fire, 1 Ice, and 1 Electric weapons"
+    },
+    {
+        "name": "Full Barbarian Set"
+    },
+    {
+        "name": "Pay 3 Great Fairies"
+    },
+    {
+        "name": "Mount the Lord of the Mountain"
+    }
+]
