+++ bingosync-app/testdata/gen_output/zelda_botw_short/1.json
@@ -1,0 +1,77 @@
+[
+    {
+        "name": "Barbarian Armor"
+    },
+    {
+        "name": "3 Akkala Shrines"
+    },
+    {
+        "name": "Any Hasty meal"
+    },
+    {
+        "name": "3 Koko Cooking Side Quests"
+    },
+    {
+        "name": "Kill a Wizzrobe of any kind"
+    },
+    {
+        "name": "'Diving is Beauty!' Side Quest"
+    },
+    {
+        "name": "Open 2 Bokoblin Camp Chests"
+    },
+    {
+        "name": "8 Shrines"
+    },
+    {
+        "name": "Kill a Frost Talus"
+    },
+    {
+        "name": "Any 'Boko' Equipment item"
+    },
+    {
+        "name": "9 Shrines"
+    },
+    {
+        "name": "Any Stalfos 'Arm' Equipment item"
+    },
+    {
+        "name": "3 Koko Cooking Side Quests"
+    },
+    {
+        "name": "6 Lake Tower Koroks"
+    },
+    {
+        "name": "Fruitcake"
+    },
+    {
+        "name": "Activate Faron Tower"
+    },
+    {
+        "name": "9 Wasteland Koroks"
+    },
+    {
+        "name": "Kill a Stone Talus (Rare)"
+    },
+    {
+        "name": "Give Beedle a Beetle"
+    },
+    {
+        "name": "Black Dyed Hylian Hood"
+    },
+    {
+        "name": "Any 'Boko' Equipment item"
+    },
+    {
+        "name": "Any Fairy Tonic"
+    },
+    {
+        "name": "Kill a fire-typed Wizzrobe"
+    },
+    {
+        "name": "Finish Golf Minigame"
+    },
+    {
+        "name": "3 Lanayru Shrines"
+    }
+]
