+++ bingosync-app/testdata/gen_output/zelda_botw_short/1000.json
@@ -1,0 +1,77 @@
+[
+    {
+        "name": "12 Dueling Peaks Koroks"
+    },
+    {
+        "name": "'The Secret Club's Secret' Side Quest"
+    },
+    {
+        "name": "Any 'Forest Dweller's' Equipment item"
+    },
+    {
+        "name": "Rock-Hard Food"
+    },
+    {
+        "name": "Hylian Hood"
+    },
+    {
+        "name": "'Balloon Flight' Side Quest"
+    },
+    {
+        "name": "Hylian Trousers"
+    },
+    {
+        "name": "Register the Giant Horse"
+    },
+    {
+        "name": "9 Wasteland Koroks"
+    },
+    {
+        "name": "3 Dueling Peaks Shrines"
+    },
+    {
+        "name": "Hylian Hood \u2605"
+    },
+    {
+        "name": "Activate Eldin Tower"
+    },
+    {
+        "name": "Kill a Wizzrobe of any kind"
+    },
+    {
+        "name": "Save a Traveler in Peril"
+    },
+    {
+        "name": "Wildberry Crepe"
+    },
+    {
+        "name": "Any Hasty meal"
+    },
+    {
+        "name": "Kill a Gaurdian Turret"
+    },
+    {
+        "name": "'Twin Memories' Shrines"
+    },
+    {
+        "name": "Activate Tabantha Tower"
+    },
+    {
+        "name": "6 Ridgeland Koroks"
+    },
+    {
+        "name": "Activate Tabantha Tower"
+    },
+    {
+        "name": "Any Stalfos 'Arm' Equipment item"
+    },
+    {
+        "name": "3 Central Tower Koroks"
+    },
+    {
+        "name": "9 Shrines"
+    },
+    {
+        "name": "Strike in Bowling Minigame"
+    }
+]
