+++ bingosync-app/testdata/gen_output/zelda_botw_short/1234.json
@@ -1,0 +1,77 @@
+[
+    {
+        "name": "Warm Doublet"
+    },
+    {
+        "name": "3 Woodland Shrines"
+    },
+    {
+        "name": "7 Side Quests"
+    },
+    {
+        "name": "Activate Akkala Tower"
+    },
+    {
+        "name": "Any Boomerang"
+    },
+    {
+        "name": "9 Tabantha Koroks"
+    },
+    {
+        "name": "Enduring Elixir"
+    },
+    {
+        "name": "Activate Lake Tower"
+    },
+    {
+        "name": "'The Silent Swordswomen' Shrine Quest"
+    },
+    {
+        "name": "Kill an ice-typed Wizzrobe"
+    },
+    {
+        "name": "3 Eldin Shrines"
+    },
+    {
+        "name": "Kill an ice-typed Wizzrobe"
+    },
+    {
+        "name": "Kill a Stone Talus (Luminous)"
+    },
+    {
+        "name": "5 Side Quests"
+    },
+    {
+        "name": "Activate Dueling Peaks Tower"
+    },
+    {
+        "name": "'Wild Horses' Side Quest"
+    },
+    {
+        "name": "5 Koroks"
+    },
+    {
+        "name": "Champion's Tunic"
+    },
+    {
+        "name": "Any 'Forest Dweller's' Equipment item"
+    },
+    {
+        "name": "8 Shrines"
+    },
+    {
+        "name": "Any Fairy Tonic"
+    },
+    {
+        "name": "Climber's Bandanna \u2605"
+    },
+    {
+        "name": "Read Paya's Diary"
+    },
+    {
+        "name": "3 Ridgeland Koroks"
+    },
+    {
+        "name": "Read Zelda's Diary"
+    }
+]
