+++ bingosync-app/testdata/gen_output/zelda_botw_short/12345.json
@@ -1,0 +1,77 @@
+[
+    {
+        "name": "Climber's Bandanna \u2605"
+    },
+    {
+        "name": "3 Identical Weapons"
+    },
+    {
+        "name": "3 Dueling Peaks Koroks"
+    },
+    {
+        "name": "Kill a Hinox of any kind"
+    },
+    {
+        "name": "'A Royal Recipe' Side Quest"
+    },
+    {
+        "name": "Open 2 Bokoblin Camp Chests"
+    },
+    {
+        "name": "'By Firefly's Light' Side Quest"
+    },
+    {
+        "name": "Finish Golf Minigame"
+    },
+    {
+        "name": "Zora Helm"
+    },
+    {
+        "name": "3 Ridgeland Koroks"
+    },
+    {
+        "name": "Activate Gerudo Tower"
+    },
+    {
+        "name": "3 Faron Koroks"
+    },
+    {
+        "name": "A Minor Test of Strength Shrine"
+    },
+    {
+        "name": "Any 'Royal Guard's' Equipment item"
+    },
+    {
+        "name": "Strike in Bowling Minigame"
+    },
+    {
+        "name": "'Secret of the Snowy Peaks' Shrine Quest"
+    },
+    {
+        "name": "Kill a Stone Talus (Luminous)"
+    },
+    {
+        "name": "Any 'Forest Dweller's' Equipment item"
+    },
+    {
+        "name": "Any Energizing meal"
+    },
+    {
+        "name": "6 Lanayru Koroks"
+    },
+    {
+        "name": "Save a Traveler in Peril"
+    },
+    {
+        "name": "Any Tough meal"
+    },
+    {
+        "name": "'Sunshroom Sensing' Side Quest"
+    },
+    {
+        "name": "8 Shrines"
+    },
+    {
+        "name": "Kill an ice-typed Wizzrobe"
+    }
+]
